Add functionality to segemented control to change language of tract pages
